+++ NEWS.md
@@ -1,3 +1,9 @@
+## jti v0.5.2.9999 (xxxx-yy-zz)
+
+ * There was a bug in the creation of the junction tree when calling Kruskals algorithm.
+ * It is now possible to specify variables of interest in advance, such that we are 
+ guaranteed to be able to query the joint pmf of these variables.
+ 
 ## jti v0.5.2 (2020-11-24)
 
  * A new way of entering evidence that is much more robust